+++ internal/pkg/instance/instance.go
@@ -2,6 +2,7 @@ package instance
 
 import (
 	"archive/zip"
+	"bufio"
 	"bytes"
 	"errors"
 	"fmt"
@@ -11,7 +12,10 @@ import (
 	"path"
 	"path/filepath"
 	"runtime"
+	"strings"
 	"time"
+
+	"golang.org/x/text/encoding/charmap"
 
 	"github.com/sirupsen/logrus"
 
@@ -37,9 +41,12 @@ type Instance struct {
 	Path   string
 	Cfg    AccWebConfigJson
 	AccCfg AccConfigFiles
+	Live   *LiveState
 
-	cmd     *exec.Cmd
-	logFile *os.File
+	logParser *logParser
+	cmd       *exec.Cmd
+	logFile   *os.File
+	cmdOut    io.ReadCloser
 }
 
 func (s *Instance) GetID() string {
@@ -51,65 +58,25 @@ func (s *Instance) Start() error {
 		return ErrServerCantBeRunning
 	}
 
-	if err := s.CheckDirectory(); err != nil {
+	if err := s.prepareInstanceDir(); err != nil {
 		return err
 	}
 
-	// Copy config files to cfg dir
-	if err := helper.CreateIfNotExists(path.Join(s.Path, accCfgDir), 0755); err != nil {
+	s.prepareCommandAndArgs()
+
+	if err := s.prepareCmdLogHandler(); err != nil {
 		return err
 	}
-
-	fileList := []string{
-		configurationJsonName,
-		settingsJsonName,
-		eventJsonName,
-		eventRulesJsonName,
-		entrylistJsonName,
-		bopJsonName,
-		assistRulesJsonName,
-	}
-
-	for _, name := range fileList {
-		if err := helper.Copy(path.Join(s.Path, name), path.Join(s.Path, accCfgDir, name)); err != nil {
-			return err
-		}
-	}
-
-	command := "." + string(filepath.Separator) + accDedicatedServerFile
-	var args []string
-
-	if runtime.GOOS == "linux" {
-		command = "wine"
-		args = []string{accDedicatedServerFile}
-	}
-
-	var err error
-	if s.logFile, err = s.createLogFile(); err != nil {
-		return err
-	}
-
-	cmd := exec.Command(command, args...)
-	cmd.Dir = s.Path
-	cmd.Stdout = s.logFile
-	cmd.Stderr = s.logFile
-
-	s.cmd = cmd
 
 	if err := s.cmd.Start(); err != nil {
 		return err
 	}
 
+	s.Live.setServerState(ServerStateStarting)
+
 	logrus.WithField("server_id", s.GetID()).WithField("pid", s.GetProcessID()).Info("acc server started")
 
-	go func(cmd *exec.Cmd, l io.Closer) {
-		// wait for shutdown or crash
-		if err := cmd.Wait(); err != nil {
-			logrus.WithError(err).Error("Error when server stopped")
-		}
-
-		_ = l.Close()
-	}(cmd, s.logFile)
+	go s.wait()
 
 	return nil
 }
@@ -124,6 +91,9 @@ func (s *Instance) Stop() error {
 			return err
 		}
 	}
+
+	s.Live.serverOffline()
+
 	logrus.WithField("server_id", s.GetID()).Info("acc server stopped")
 
 	return nil
@@ -288,4 +258,91 @@ func (s *Instance) createLogFile() (*os.File, error) {
 	filename := fmt.Sprintf("logs_%s_%s%s", time.Now().Format(logTimeFormat), s.GetID(), logExt)
 
 	return os.Create(path.Join(s.Path, logDir, filename))
+}
+
+func (s *Instance) prepareInstanceDir() error {
+	if err := s.CheckDirectory(); err != nil {
+		return err
+	}
+
+	// Copy config files to cfg dir
+	if err := helper.CreateIfNotExists(path.Join(s.Path, accCfgDir), 0755); err != nil {
+		return err
+	}
+
+	fileList := []string{
+		configurationJsonName,
+		settingsJsonName,
+		eventJsonName,
+		eventRulesJsonName,
+		entrylistJsonName,
+		bopJsonName,
+		assistRulesJsonName,
+	}
+
+	for _, name := range fileList {
+		if err := helper.Copy(path.Join(s.Path, name), path.Join(s.Path, accCfgDir, name)); err != nil {
+			return err
+		}
+	}
+
+	return nil
+}
+
+func (s *Instance) wait() {
+	// wait for shutdown or crash
+	if err := s.cmd.Wait(); err != nil {
+		logrus.WithError(err).Error("Error when server stopped")
+	}
+
+	_ = s.cmdOut.Close()
+	_ = s.logFile.Close()
+}
+
+func (s *Instance) prepareCommandAndArgs() {
+	command := "." + string(filepath.Separator) + accDedicatedServerFile
+	var args []string
+
+	if runtime.GOOS == "linux" {
+		command = "wine"
+		args = []string{accDedicatedServerFile}
+	}
+
+	cmd := exec.Command(command, args...)
+	cmd.Dir = s.Path
+	s.cmd = cmd
+}
+
+func (s *Instance) prepareCmdLogHandler() error {
+	if s.cmd == nil {
+		return errors.New("instance command not prepared")
+	}
+
+	var err error
+	s.cmdOut, err = s.cmd.StdoutPipe()
+	if err != nil {
+		return err
+	}
+
+	if s.logFile, err = s.createLogFile(); err != nil {
+		return err
+	}
+
+	s.logParser = newLogParser()
+
+	r := charmap.ISO8859_1.NewDecoder().Reader(io.TeeReader(s.cmdOut, s.logFile))
+	scanner := bufio.NewScanner(r)
+
+	go func() {
+		for scanner.Scan() {
+			data := scanner.Text()
+			s.logParser.processLine(s.Live, strings.TrimSpace(data))
+		}
+
+		if err := scanner.Err(); err != nil {
+			logrus.Infof("Error while reading server console: %v", err)
+		}
+	}()
+
+	return nil
 }